Created an example to check lazy load working. A laptop has multiple outputports. after creating the laptop and the output ports in the database, close the session. If we dont close the session, hibernate will not fire the select query as the session.get() will try to get the object from the existing session memory itself.
After closing the session, create a new session and get the laptop and
print just the laptop details. You can see in the log that only
select-laptop is run without the o/p ports as the OneToMany relationship
is lazy loaded by default.

Next running the getOutputPorts on the laptop instance will fire the
select-outputports on the backend.
This proves that it is lazyloaded